Bugfix: libnuma was failing RISC-V
On RISC-V libnuma is available, but does not seem to be fully
functional.  Thus, some of the call reported undefined values and
produced an error when the runtime tried to detect the NUMA structure
of the RISC-V system.  Thus bugfix changes the detection code such
that it first determines if libnuma claims to work and if not falls
back to a single NUMA domain (like with systems w/o libnuma).